+++ CHANGES.md
@@ -1,5 +1,12 @@
 ## Changelog ##
 
+### 1.2 ###
+* Added: VisualComponents Extension to visualize the HTML Element Structure in Visual Editor Mode (headings,section,p)
+* Added: Option to remove Rewrite Rules for all types of Feeds (RSS, RTF, ATOM, RSS2)
+* Changed: "Disable XMLRPC" will now force the `xmlrpc.php` Endpoint to return an HTTP403 Response
+* Changed: TinyMCE Autowidth is set to **95%**
+* Changed: **Disable oEmbeds** removes also the related Rewrite Rules as well as the endpoint (embed.php template page)
+
 ### 1.1 ###
 * Added: Google Analytics Tracking Option
 * Added: Piwik Analaytics Tracking Option